Arreglo escritura ppf y potencia actual en pantallas del display_dogs164
El error estaba en que no se limpiaban los valores viejos cuando si iban
a volver a escribir entonces quedaba cualquier cosa.

+++ indoor_control/include/global_manager.h
@@ -38,7 +38,7 @@ typedef struct
     flora_vege_status_t flora_vege_status;
     pwm_auto_info_t pwm_auto;
     s_out_auto_info_t s_out_auto;
-    uint16_t ppf;
+    uint32_t ppf;
 } global_manager_nv_t;
 
 typedef struct
@@ -121,8 +121,8 @@ uint8_t global_manager_get_s_out_turn_on_time(struct tm *turn_on_time, uint8_t s
 uint8_t global_manager_set_s_out_time_enable_status(uint8_t time_enable_status, uint8_t s_out_index);
 uint8_t global_manager_get_s_out_time_enable_status(uint8_t *time_enable_status, uint8_t s_out_index);
 
-uint8_t global_manager_get_ppf(uint16_t *ppf);
-uint8_t global_manager_set_ppf(uint16_t ppf);
+uint8_t global_manager_get_ppf(uint32_t *ppf);
+uint8_t global_manager_set_ppf(uint32_t ppf);
 
 /*
     UPDATE PWM CALENDAR
